Chore: Inital refactor
